remove old admin code - complete cleanup
consolidated admin to single inertia-based system. deleted legacy reactonrails components, orphaned erb views, and duplicate packs.

changes:
- deleted old admin.ts pack (reactonrails)
- renamed admin_inertia.ts → admin.ts
- removed url interceptor and moved clipboard handlers
- deleted admin_old.html.erb layout
- removed controller render override
- deleted 7 orphaned erb views (controllers use inertia without legacy_template)
- deleted 3 unused server-components (Nav, SearchPopover, SalesReportsPage)
- updated ssr.ts to remove deleted components
- fixed links_controller params.permit error (unrelated but blocking)

admin now fully on inertia with no legacy code. closes #2060
