--- app/javascript/components/server-components/Admin/SalesReportsPage.tsx
@@ -1,143 +1,0 @@
-import * as React from "react";
-import { createCast } from "ts-safe-cast";
-
-import { register } from "$app/utils/serverComponentUtil";
-
-import { Form } from "$app/components/server-components/Admin/Form";
-import { showAlert } from "$app/components/server-components/Alert";
-import Placeholder from "$app/components/ui/Placeholder";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "$app/components/ui/Table";
-
-type JobHistoryItem = {
-  job_id: string;
-  country_code: string;
-  start_date: string;
-  end_date: string;
-  sales_type: string;
-  enqueued_at: string;
-  status: string;
-  download_url?: string;
-};
-
-type Props = {
-  countries: [string, string][];
-  sales_types: [string, string][];
-  job_history: JobHistoryItem[];
-  form_action: string;
-  authenticity_token: string;
-};
-
-const AdminSalesReportsPage = ({ countries, sales_types, job_history, form_action, authenticity_token }: Props) => {
-  const countryCodeToName = React.useMemo(() => {
-    const map: Record<string, string> = {};
-    countries.forEach(([name, code]) => {
-      map[code] = name;
-    });
-    return map;
-  }, [countries]);
-
-  const salesTypeCodeToName = React.useMemo(() => {
-    const map: Record<string, string> = {};
-    sales_types.forEach(([code, name]) => {
-      map[code] = name;
-    });
-    return map;
-  }, [sales_types]);
-
-  return (
-    <>
-      <Form
-        url={form_action}
-        method="POST"
-        confirmMessage={false}
-        onSuccess={() => {
-          showAlert("Sales report job enqueued successfully!", "success");
-          window.location.reload();
-        }}
-      >
-        {(isLoading) => (
-          <section>
-            <header>Generate sales report with custom date ranges</header>
-
-            <label htmlFor="country_code">Country</label>
-            <select name="sales_report[country_code]" id="country_code" required>
-              <option value="">Select country</option>
-              {countries.map(([name, code]) => (
-                <option key={code} value={code}>
-                  {name}
-                </option>
-              ))}
-            </select>
-
-            <label htmlFor="start_date">Start date</label>
-            <input name="sales_report[start_date]" id="start_date" type="date" required />
-
-            <label htmlFor="end_date">End date</label>
-            <input name="sales_report[end_date]" id="end_date" type="date" required />
-
-            <label htmlFor="sales_type">Type of sales</label>
-            <select name="sales_report[sales_type]" id="sales_type" required>
-              <option value="">Select sales type</option>
-              {sales_types.map(([code, name]) => (
-                <option key={code} value={code}>
-                  {name}
-                </option>
-              ))}
-            </select>
-
-            <button type="submit" className="button primary" disabled={isLoading}>
-              {isLoading ? "Generating..." : "Generate report"}
-            </button>
-
-            <input type="hidden" name="authenticity_token" value={authenticity_token} />
-          </section>
-        )}
-      </Form>
-
-      <section>
-        {job_history.length > 0 ? (
-          <Table>
-            <TableHeader>
-              <TableRow>
-                <TableHead>Country</TableHead>
-                <TableHead>Date range</TableHead>
-                <TableHead>Sales Type</TableHead>
-                <TableHead>Enqueued at</TableHead>
-                <TableHead>Status</TableHead>
-                <TableHead>Download</TableHead>
-              </TableRow>
-            </TableHeader>
-            <TableBody>
-              {job_history.map((job, index) => (
-                <TableRow key={index}>
-                  <TableCell>{countryCodeToName[job.country_code] || job.country_code}</TableCell>
-                  <TableCell>
-                    {job.start_date} to {job.end_date}
-                  </TableCell>
-                  <TableCell>{job.sales_type ? salesTypeCodeToName[job.sales_type] : sales_types[0]?.[1]}</TableCell>
-                  <TableCell>{new Date(job.enqueued_at).toLocaleString()}</TableCell>
-                  <TableCell>{job.status}</TableCell>
-                  <TableCell>
-                    {job.status === "completed" && job.download_url ? (
-                      <a href={job.download_url} className="button small" target="_blank" rel="noopener noreferrer">
-                        Download CSV
-                      </a>
-                    ) : (
-                      <span>-</span>
-                    )}
-                  </TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        ) : (
-          <Placeholder>
-            <h2>No sales reports generated yet.</h2>
-          </Placeholder>
-        )}
-      </section>
-    </>
-  );
-};
-
-export default register({ component: AdminSalesReportsPage, propParser: createCast() });
