+++ app/javascript/ssr.ts
@@ -10,14 +10,11 @@ import AdminChangeEmailForm from "$app/components/server-components/Admin/Change
 import AdminFlagForFraudForm from "$app/components/server-components/Admin/FlagForFraudForm";
 import AdminManualPayoutForm from "$app/components/server-components/Admin/ManualPayoutForm";
 import AdminMassTransferPurchasesForm from "$app/components/server-components/Admin/MassTransferPurchasesForm";
-import AdminNav from "$app/components/server-components/Admin/Nav";
 import AdminPausePayoutsForm from "$app/components/server-components/Admin/PausePayoutsForm";
 import AdminProductAttributesAndInfo from "$app/components/server-components/Admin/ProductAttributesAndInfo";
 import AdminProductPurchases from "$app/components/server-components/Admin/ProductPurchases";
 import AdminProductStats from "$app/components/server-components/Admin/ProductStats";
 import AdminResendReceiptForm from "$app/components/server-components/Admin/ResendReceiptForm";
-import AdminSalesReportsPage from "$app/components/server-components/Admin/SalesReportsPage";
-import AdminSearchPopover from "$app/components/server-components/Admin/SearchPopover";
 import AdminSetCustomFeeForm from "$app/components/server-components/Admin/SetCustomFeeForm";
 import AdminSuspendForFraudForm from "$app/components/server-components/Admin/SuspendForFraudForm";
 import AdminSuspendForTosForm from "$app/components/server-components/Admin/SuspendForTosForm";
@@ -99,14 +96,11 @@ ReactOnRails.register({
   AdminFlagForFraudForm,
   AdminManualPayoutForm,
   AdminMassTransferPurchasesForm,
-  AdminNav,
   AdminPausePayoutsForm,
   AdminProductAttributesAndInfo,
   AdminProductPurchases,
   AdminProductStats,
-  AdminSalesReportsPage,
   AdminResendReceiptForm,
-  AdminSearchPopover,
   AdminSetCustomFeeForm,
   AdminSuspendForFraudForm,
   AdminSuspendForTosForm,
