--- app/views/admin/base/index.html.erb
@@ -1,24 +1,0 @@
-<%= load_pack("admin") %>
-
-<section>
-  <div class="flex flex-col gap-4 max-w-2xl">
-    <%= form_tag "#", class: "flex flex-col gap-4" do %>
-      <%= text_field_tag :user_identifier, nil,
-          placeholder: "Enter user email, username, or Stripe account ID",
-          class: "w-full rounded-md", autofocus: true %>
-
-      <div class="flex gap-4">
-        <%= button_tag "Impersonate user",
-            type: "submit",
-            formaction: admin_impersonate_path,
-            formmethod: "get",
-            class: "button w-auto flex-1" %>
-        <%= button_tag "View Stripe account",
-            type: "submit",
-            formaction: admin_redirect_to_stripe_dashboard_path,
-            formmethod: "get",
-            class: "button w-auto flex-1" %>
-      </div>
-    <% end %>
-  </div>
-</section>
