--- app/views/admin/block_email_domains/show.html.erb
@@ -1,2 +1,0 @@
-<%= load_pack("admin") %>
-<%= render partial: "admin/block_email_domains/form", locals: { url: admin_block_email_domains_path, header: "To suspend email domains, please enter them separated by comma or newline.", button_label: "Block email domains" } %>
