--- app/views/admin/merchant_accounts/show.html.erb
@@ -1,5 +1,0 @@
-<%= load_pack("admin") %>
-
-<div>
-  <%= render("admin/merchant_accounts/merchant_account", merchant_account: @merchant_account) %>
-</div>
