--- app/views/admin/sales_reports/index.html.erb
@@ -1,12 +1,0 @@
-<%= load_pack("admin") %>
-
-<div>
-  <%= react_component "Alert", props: { initial: nil }, prerender: true %>
-
-  <%= react_component "AdminSalesReportsPage", props: {
-    countries: @countries,
-    job_history: @job_history,
-    form_action: admin_sales_reports_path,
-    authenticity_token: form_authenticity_token,
-  } %>
-</div>
