--- app/views/admin/unblock_email_domains/show.html.erb
@@ -1,2 +1,0 @@
-<%= load_pack("admin") %>
-<%= render partial: "admin/block_email_domains/form", locals: { url: admin_unblock_email_domains_path, header: "To unsuspend email domains, please enter them separated by comma or newline.", button_label: "Unblock email domains" } %>
