+++ app/views/layouts/admin.html.erb
@@ -11,7 +11,7 @@
     <%= yield %>
 
     <%= javascript_include_tag "application" %>
-    <%= load_pack("admin_inertia") %>
+    <%= load_pack("admin") %>
     <%= yield :script %>
   <% end %>
 </html>
